Refactored to get VideoTiles
